requirements.txt
fastapi
sqlmodel
psycopg[binary]
python-dotenv
